pyproject.toml
[project]
name = "aseprite-mcp"
version = "0.1.0"
description = "MCP server for creating pixel art with Aseprite through AI assistants"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
]

[project.optional-dependencies]
analysis = [
    "Pillow>=10.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aseprite_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
